requirements.txt
Click
python-dotenv
tweepy==3.9.0
geoalchemy2==0.8.4
sqlalchemy==1.3.18
sqlalchemy_utils==0.36.8
pycountry==20.7.3
nltk==3.5
spacy==2.3.2
pymorphy2==0.8
pymorphy2-dicts
descartes==1.1.0
tika==1.24
wordcloud==1.8.1
pandas==1.1.3
psycopg2-binary